# Add type header to WAL/Checkpoint records and use TSDB records