**What this PR does**:

Add a single-byte type header to WAL and Checkpoint records and way to handle old records concurrently. Additionally, replace existing proto records in WAL with Prometheus TSDB records. 

**Which issue(s) this PR fixes**:
Fixes #2426

**Checklist**
- [x] Tests updated
- [ ] Documentation added
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`