@codesome 2020-04-14
After some offline discussion with Goutham, we decided to add the Prometheus TSDB records in this PR itself to not have many versions or records, as the current record would be obsolete immediately after TSDB records are in.