@codesome 2020-04-15
I wrote a benchmark for WAL replay (will open PR after https://github.com/cortexproject/cortex/pull/2465 is merged as it depends on that) to compare the new records, and the results look good. Largely depends on Checkpoint/WAL ratio. The benchmark had equal number of samples in WAL and Checkpoint.
```
benchmark                    old ns/op      new ns/op      delta
BenchmarkWALReplay/#00-8     2027607295     1656442045     -18.31%

benchmark                    old allocs     new allocs     delta
BenchmarkWALReplay/#00-8     3050719        1142134        -62.56%

benchmark                    old bytes      new bytes      delta
BenchmarkWALReplay/#00-8     5943738312     5638734152     -5.13%
```